+++ src/config/serialization.rs
@@ -201,6 +201,7 @@ mod tests {
         assert_eq!(cfg.user_interface.text_editor, PathBuf::from("vim"));
         assert_eq!(cfg.user_interface.show_hidden_files, true);
         assert_eq!(cfg.user_interface.check_for_updates.unwrap(), true);
+        assert_eq!(cfg.user_interface.prompt_on_file_replace.unwrap(), false);
         assert_eq!(cfg.user_interface.group_dirs, Some(String::from("last")));
         assert_eq!(
             cfg.user_interface.file_fmt,
@@ -244,6 +245,7 @@ mod tests {
         assert_eq!(cfg.user_interface.show_hidden_files, true);
         assert_eq!(cfg.user_interface.group_dirs, None);
         assert!(cfg.user_interface.check_for_updates.is_none());
+        assert!(cfg.user_interface.prompt_on_file_replace.is_none());
         assert!(cfg.user_interface.file_fmt.is_none());
         assert!(cfg.user_interface.remote_file_fmt.is_none());
         // Verify keys
@@ -317,6 +319,7 @@ mod tests {
         text_editor = "vim"
         show_hidden_files = true
         check_for_updates = true
+        prompt_on_file_replace = false
         group_dirs = "last"
         file_fmt = "{NAME} {PEX}"
         remote_file_fmt = "{NAME} {USER}"