+++ src/ui/activities/filetransfer/update.rs
@@ -27,14 +27,14 @@
  */
 // locals
 use super::{
-    actions::SelectedEntry, browser::FileExplorerTab, FileTransferActivity, LogLevel,
+    actions::SelectedEntry, browser::FileExplorerTab, FileTransferActivity, LogLevel, TransferOpts,
     COMPONENT_EXPLORER_FIND, COMPONENT_EXPLORER_LOCAL, COMPONENT_EXPLORER_REMOTE,
     COMPONENT_INPUT_COPY, COMPONENT_INPUT_EXEC, COMPONENT_INPUT_FIND, COMPONENT_INPUT_GOTO,
     COMPONENT_INPUT_MKDIR, COMPONENT_INPUT_NEWFILE, COMPONENT_INPUT_OPEN_WITH,
     COMPONENT_INPUT_RENAME, COMPONENT_INPUT_SAVEAS, COMPONENT_LIST_FILEINFO, COMPONENT_LOG_BOX,
     COMPONENT_PROGRESS_BAR_FULL, COMPONENT_PROGRESS_BAR_PARTIAL, COMPONENT_RADIO_DELETE,
-    COMPONENT_RADIO_DISCONNECT, COMPONENT_RADIO_QUIT, COMPONENT_RADIO_SORTING,
-    COMPONENT_TEXT_ERROR, COMPONENT_TEXT_FATAL, COMPONENT_TEXT_HELP,
+    COMPONENT_RADIO_DISCONNECT, COMPONENT_RADIO_QUIT, COMPONENT_RADIO_REPLACE,
+    COMPONENT_RADIO_SORTING, COMPONENT_TEXT_ERROR, COMPONENT_TEXT_FATAL, COMPONENT_TEXT_HELP,
 };
 use crate::fs::explorer::FileSorting;
 use crate::fs::FsEntry;
@@ -358,7 +358,7 @@ impl Update for FileTransferActivity {
                 }
                 (COMPONENT_EXPLORER_FIND, key) if key == &MSG_KEY_SPACE => {
                     // Get entry
-                    self.action_find_transfer(None);
+                    self.action_find_transfer(TransferOpts::default());
                     // Reload files
                     match self.browser.tab() {
                         // NOTE: swapped by purpose
@@ -583,7 +583,7 @@ impl Update for FileTransferActivity {
                         FileExplorerTab::Remote => self.action_remote_saveas(input.to_string()),
                         FileExplorerTab::FindLocal | FileExplorerTab::FindRemote => {
                             // Get entry
-                            self.action_find_transfer(Some(input.to_string()));
+                            self.action_find_transfer(TransferOpts::default().save_as(input));
                         }
                     }
                     self.umount_saveas();
@@ -653,6 +653,21 @@ impl Update for FileTransferActivity {
                     }
                 }
                 (COMPONENT_RADIO_DELETE, _) => None,
+                // -- replace
+                (COMPONENT_RADIO_REPLACE, key)
+                    if key == &MSG_KEY_ESC
+                        || key == &Msg::OnSubmit(Payload::One(Value::Usize(1))) =>
+                {
+                    self.umount_radio_replace();
+                    None
+                }
+                (COMPONENT_RADIO_REPLACE, Msg::OnSubmit(Payload::One(Value::Usize(0)))) => {
+                    // Choice is 'YES'
+                    self.umount_radio_replace();
+                    self.action_finalize_pending_transfer();
+                    None
+                }
+                (COMPONENT_RADIO_REPLACE, _) => None,
                 // -- disconnect
                 (COMPONENT_RADIO_DISCONNECT, key)
                     if key == &MSG_KEY_ESC